+ new file structure inorder to have a testing enironnement
