+++ package.json
@@ -27,15 +27,18 @@
         "@types/dotenv": "^8.2.0",
         "@types/express": "^4.17.11",
         "@types/jest": "^26.0.24",
+        "@types/mongodb-memory-server": "^2.3.0",
         "@types/morgan": "^1.9.2",
         "@typescript-eslint/eslint-plugin": "^4.22.0",
         "@typescript-eslint/parser": "^4.22.0",
+        "apollo-server-testing": "^2.25.2",
         "eslint": "^5.16.0",
         "eslint-config-airbnb-base": "^14.2.1",
         "eslint-config-prettier": "^8.3.0",
         "eslint-plugin-import": "^2.22.1",
         "eslint-plugin-prettier": "^3.4.0",
         "jest": "^27.0.6",
+        "mongodb-memory-server": "^7.3.4",
         "prettier": "^2.2.1",
         "ts-jest": "^27.0.4",
         "ts-node": "^9.1.1",
